Drop unneeded interface methods
`AccessTokenEntityInterface::getScopes()` is not needed as the interface
extends `TokenInterface` which declares the identical method (including
identical PHPDoc).

The method should be dropped as well to ease compatibility with
`league/oauth2-server` 8 & 9 as the method of the base interface adds
the `array` return type, which is harder to add to
`AccessTokenEntityInterface` while maintaining compatibility with
`league/oauth2-server` 8 (as the `getScopes()` implementation in
`AccessTokenTrait` isn't compatible because that still doesn't have a
return type). While when dropping the method declaration here makes sure
that the interface is declared in `league/oauth2-server` and the actual
implementation of the interface method is also in `league/oauth2-server`
which means it's always compatible.